ray tracing legacy made
