+++ autolens/aggregator/fit_imaging.py
@@ -19,7 +19,6 @@ def _fit_imaging_from(
     settings_pixelization: aa.SettingsPixelization = None,
     settings_inversion: aa.SettingsInversion = None,
     use_preloaded_grid: bool = True,
-    use_hyper_scaling: bool = True,
 ) -> FitImaging:
     """
     Returns a `FitImaging` object from a PyAutoFit database `Fit` object and an instance of galaxies from a non-linear
@@ -75,7 +74,6 @@ def _fit_imaging_from(
         settings_pixelization=settings_pixelization,
         settings_inversion=settings_inversion,
         preloads=preloads,
-        use_hyper_scaling=use_hyper_scaling,
     )
 
 
@@ -87,7 +85,6 @@ class FitImagingAgg(AbstractAgg):
         settings_pixelization: Optional[aa.SettingsPixelization] = None,
         settings_inversion: Optional[aa.SettingsInversion] = None,
         use_preloaded_grid: bool = True,
-        use_hyper_scaling: bool = True,
     ):
         """
         Wraps a PyAutoFit aggregator in order to create generators of fits to imaging data, corresponding to the
@@ -99,7 +96,6 @@ class FitImagingAgg(AbstractAgg):
         self.settings_pixelization = settings_pixelization
         self.settings_inversion = settings_inversion
         self.use_preloaded_grid = use_preloaded_grid
-        self.use_hyper_scaling = use_hyper_scaling
 
     def make_object_for_gen(self, fit, galaxies) -> FitImaging:
         """
@@ -126,5 +122,4 @@ class FitImagingAgg(AbstractAgg):
             settings_pixelization=self.settings_pixelization,
             settings_inversion=self.settings_inversion,
             use_preloaded_grid=self.use_preloaded_grid,
-            use_hyper_scaling=self.use_hyper_scaling,
         )
