+++ autolens/legacy/__init__.py
@@ -1,1 +1,3 @@
+from autogalaxy.legacy import *
+from autolens.legacy.lens.ray_tracing import Tracer
 from autolens.legacy.imaging.fit_imaging import FitImaging
